update cors and endpoints

+++ src/main/java/com/example/excercise2/controller/EditorController.java
@@ -3,15 +3,21 @@ package com.example.excercise2.controller;
 import com.example.excercise2.entity.Editor;
 import com.example.excercise2.repositories.EditorRepository;
 import com.example.excercise2.service.EditorService;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.Objects;
 
 @RestController
 @CrossOrigin
-@RequestMapping("editor")
-public class EditorController {
+public class EditorController extends BaseController {
     final EditorRepository editorRepository;
     final EditorService editorService;
 
@@ -20,38 +26,49 @@ public class EditorController {
         this.editorService = editorService;
     }
 
-    @GetMapping("/all")
-    public List<Editor> getAllEditors() {
-        return editorRepository.findAll();
+    @GetMapping("/editors")
+    public List<Editor> getEditors() {
+        return editorService.getEditorsForUser(getCurrentUser().getUserId());
     }
 
-    @GetMapping("/editors/{userId}")
-    public List<Editor> getEditorsByUserId(@PathVariable("userId") String userId) {
-        return editorRepository.findEditorsByUserId(userId);
-    }
-
-    @GetMapping(value ="/{editorId}")
-    public Editor getEditorById(@PathVariable("editorId") String editorId) {
-        Editor editor = editorRepository.findEditorById(editorId);
-        if(Objects.nonNull(editor)){
-            return editor;
-        }
-        throw new RuntimeException("Editor not found");
-    }
-
-    @PostMapping("/create")
+    @PostMapping("/editors")
     public Editor createEditor(@RequestBody Editor editor) {
+        editor.setUserId(getCurrentUser().getUserId());
         return editorService.createEditor(editor);
     }
 
-    @PatchMapping("/update/{editorId}")
-    public Editor updateEditor(@PathVariable("editorId") String editorId, @RequestBody Editor editor) {
-        return editorService.updateEditor(editorId, editor);
+    @GetMapping(value = "/editor/{editorId}")
+    public Editor getEditorById(@PathVariable("editorId") String editorId) {
+        Editor editor = editorService.getEditorById(editorId);
+        if (Objects.nonNull(editor)) {
+            if (editor.getPublic()) {
+                return editor;
+            } else {
+                if (editor.getUserId().equals(getCurrentUser().getUserId())) {
+                    return editor;
+                }
+            }
+        }
+        throw new RuntimeException("Editor not found!");
     }
 
-    @DeleteMapping("/delete/{editorId}")
+    @PutMapping("/editor/{editorId}")
+    public Editor updateEditor(@PathVariable("editorId") String editorId, @RequestBody Editor editor) {
+        Editor existing = editorService.getEditorById(editorId);
+        if (Objects.nonNull(existing) &&
+            (existing.getPublic() || (existing.getUserId().equals(getCurrentUser().getUserId())))) {
+            return editorService.updateEditor(editorId, editor);
+        }
+        throw new RuntimeException("Editor not found!");
+    }
+
+    @DeleteMapping("/editor/{editorId}")
     public String deleteEditor(@PathVariable("editorId") String editorId) {
-        return editorService.deleteEditor(editorId);
+        Editor existing = editorService.getEditorById(editorId);
+        if (Objects.nonNull(existing) && existing.getUserId().equals(getCurrentUser().getUserId())) {
+            return editorService.deleteEditor(editorId);
+        }
+        throw new RuntimeException("Editor not found!");
     }
